Create readme_for_basic_py.txt
Contributing:

Learning Python is an exciting adventure, and your contributions to this repository can make that journey more enjoyable for others. Whether you're new to programming or a seasoned developer, your insights are valuable in helping beginners grasp the essentials of Python.

Join us in exploring the basics of Python programming and building a supportive community of learners.

Happy coding!